# docs/swagger.yaml
definitions:
...
  models.Req:
    properties:
      filter:
        additionalProperties: true
        description: "Filter is the query object for building the value for the `WHERE`
          clause.\nThe full documentation of the supported operators is writtern in
          the README.\nAn example for filter object:\n\n\tparams, err := p.Parse([]byte(`{\n\t\t\"filter\":
          {\n\t\t\t\"account\": { \"$like\": \"%github%\" },\n\t\t\t\"$or\": [\n\t\t\t\t{
          \"city\": \"TLV\" },\n\t\t\t\t{ \"city\": \"NYC\" }\n\t\t\t]\n\t\t}\n\t}`))"
        type: object
      limit:
        description: Limit must be > 0 and <= to `LimitMaxValue`.
        type: integer
      offset:
        description: Offset must be >= 0.
        type: integer
      select:
        description: "Select contains the list of expressions define the value for
          the `SELECT` clause.\nFor example:\n\n\tparams, err := p.Parse([]byte(`{\n\t\t\"select\":
          [\"name\", \"age\"]\n\t}`))"
        items:
          type: string
        type: array
      sort:
        description: "Sort contains list of expressions define the value for the `ORDER
          BY` clause.\nIn order to return the rows in descending order you can prefix
          your field with `-`.\nFor example:\n\n\tparams, err := p.Parse([]byte(`{\n\t\t\"sort\":
          [\"name\", \"-age\", \"+redundant\"]\n\t}`))"
        items:
          type: string
        type: array
    type: object
...
  models.Resp:
    properties:
      code:
        description: 错误代码
        type: integer
      data:
        description: 数据内容
      msg:
        description: 消息提示
        type: string
      request_id:
        description: 请求ID
        type: string
    type: object
...
info:
  contact:
    email: linchao@mucfc.com
    name: 林超
    url: https://ops.mucfc.com
  description: uWork部署作业API，支持腾讯云/阿里云API操作、K8S集群API、Helm部署操作API
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: https://ops.mucfc.com
  title: uWork部署作业API
  version: 1.0.0
paths:
...
  /api/v1/router/list:
    post:
      consumes:
      - application/x-json-stream
      operationId: GetRouters
      parameters:
      - description: RQL查询json
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.Req'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Resp'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Resp'
      security:
      - ApiKeyAuth: []
      summary: '[系统内部]获取系统路由'
      tags:
      - '[系统内部]路由'
...
  /heatch_check:
    get:
      consumes:
      - application/x-json-stream
      operationId: HeathCheck
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Resp'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Resp'
      summary: '[系统内部]健康检查接口'
      tags:
      - '[系统内部]路由'
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"